+++ src/Polynomials/CompWiseTensorPolyBases.jl
@@ -59,12 +59,15 @@ struct CompWiseTensorPolyBasis{D,V,PT} <: PolynomialBasis{D,V,PT}
     @check size(orders,2) == D msg3
     @check isconcretetype(PT) "PT needs to be a concrete <:Polynomial type"
 
+    VV = ifelse(isconcretetype(V),V,typeof(zero(V)))
+    @check isconcretetype(VV) "Trying to build CompWiseTensorPolyBasis with non-concrete output type $(VV)"
+
     K = maximum(orders)
     num_poly = mapreduce(length, +, comp_terms)
 
     #TODO check orders in `orders` greater or equal than max index in terms
 
-    new{D,V,PT}(num_poly,K,orders,comp_terms)
+    new{D,VV,PT}(num_poly,K,orders,comp_terms)
   end
 end